// src/redux/types/actionTypes.ts
export enum appActionTypes {
    SET_ERROR = 'SET_ERROR',
    SET_STATUS = 'SET_STATUS',
}

export enum authActionTypes {
    REGISTER = 'REGISTER',
    LOGIN = 'LOGIN',
    LOG_OUT = '  LOG_OUT',
    SET_UID = 'SET_UID',
    SET_USER_DATA = 'SET_USER_DATA',
}

export enum galleryActionTypes {
    SET_USERS_EMAIL = ' SET_USERS_EMAIL',
    SET_SELECTED_USER = ' SET_SELECTED_USER',
    SET_IMAGES = ' SET_IMAGES',
    GET_USER_EMAILS = 'GET_USER_EMAILS',
    SAVE_ART = ' SAVE_ART',
    GET_ART = 'GET_ART',
}